Feat: 학습률 업데이트 기능 Service 구현

+++ user-api/src/main/java/com/biengual/userapi/learning/presentation/dto/UpdateLearningRateDto.java
@@ -3,7 +3,7 @@ package com.biengual.userapi.learning.presentation.dto;
 public class UpdateLearningRateDto {
     public record Request(
        Long contentId,
-       Long learningRate
+       Integer learningRate
     ) {
     }
 }
